+++ backend/src/main/java/com/bakdata/conquery/apiv1/query/ResultHeaders.java
@@ -1,4 +1,4 @@
-package com.bakdata.conquery;
+package com.bakdata.conquery.apiv1.query;
 
 import java.util.Set;
